Fix a warning in the 32-bit build.

+++ src/elf.cc
@@ -749,7 +749,7 @@ static uint64_t ToVMAddr(size_t addr, uint64_t ndx, bool is_object) {
     if (ndx >= 1 << 24) {
       THROW("ndx overflow: too many sections");
     }
-    if (addr >= 1ULL << 40) {
+    if (addr >= ((uint64_t)1) << 40) {
       THROW("address overflow: section too big");
     }
     return (ndx << 40) | addr;
